+++ backends/default/xla.go
@@ -1,0 +1,5 @@
+//go:build !noxla
+
+package _default
+
+import _ "github.com/gomlx/gomlx/backends/xla"